+++ src/blocks/InternalLink/InternalLink.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from 'vitest';
-import { renderToFragment } from '@lib/renderer/renderer';
+import { renderToFragment } from '@lib/renderer';
 import type { InternalLinkFragment } from '@lib/datocms/datocms';
 import InternalLink, { type Props as InternalLinkProps } from './InternalLink.astro';
