+++ src/blocks/TextBlock/TextBlock.test.ts
@@ -1,4 +1,4 @@
-import { renderToFragment } from '@lib/renderer/renderer';
+import { renderToFragment } from '@lib/renderer';
 import { describe, expect, test } from 'vitest';
 import TextBlock, { type Props } from './TextBlock.astro';
