+++ src/blocks/TextImageBlock/TextImageBlock.test.ts
@@ -1,4 +1,4 @@
-import { renderToFragment } from '@lib/renderer/renderer';
+import { renderToFragment } from '@lib/renderer';
 import type { TextImageBlockFragment } from '@lib/datocms/datocms';
 import { describe, expect, test } from 'vitest';
 import TextImageBlock from './TextImageBlock.astro';
