+++ src/components/Accordion/AccordionItem.test.ts
@@ -1,4 +1,4 @@
-import { renderToFragment } from '@lib/renderer/renderer';
+import { renderToFragment } from '@lib/renderer';
 import { describe, expect, test } from 'vitest';
 import AccordionItem, { type Props } from './AccordionItem.astro';
